Share the archive insert column list and parameter binder
MysqlArchiveOperations and PostgresqlArchiveOperations each carried the 31-column
archive list, its value placeholders, and a 31-parameter binder that bound the
columns in identical schema order. Only the four UUID columns differed: MySQL
encodes them as BINARY(16), PostgreSQL stores native uuid. Adding or reordering a
column meant editing both files in lockstep, with a silent column/value mismatch
on any miss.

Move the column list, the placeholders, and the binder into
ArchiveParameterBinder in store-core. The binder takes a Function<UUID, Object> so
each store supplies its own id encoding (UuidByteArrayConverter::toBytes for
MySQL, identity for PostgreSQL); every other column binds the same way for both.
