+++ ratchet-store-mysql/src/main/java/run/ratchet/store/mysql/MysqlArchiveOperations.java
@@ -32,25 +32,12 @@ import run.ratchet.store.id.UuidV7Factory;
 import run.ratchet.store.mysql.converter.UuidByteArrayConverter;
 import run.ratchet.store.spi.ArchiveStore;
 import run.ratchet.store.util.ArchiveHelper;
+import run.ratchet.store.util.ArchiveParameterBinder;
 import run.ratchet.store.util.ArchiveQuerySupport;
 import run.ratchet.store.util.ArchiveRowMapper;
 import run.ratchet.store.util.RowValues;
 
 final class MysqlArchiveOperations implements ArchiveStore {
-
-  private static final String ARCHIVE_COLUMNS =
-      """
-      archive_id, original_job_id, final_status, job_type, priority, total_attempts,
-      max_retries, backoff_policy, backoff_param_ms, timeout_sec, target_class,
-      method_name, business_key, cron_expr, zone_id, original_scheduled_time,
-      original_created_at, first_execution_time, completion_time,
-      total_execution_time_ms, queue_wait_ms, archived_at, archived_by, archive_reason,
-      job_result, result_type, final_error, payload_summary, depended_on, superseded_by,
-      tags
-      """;
-
-  private static final String ARCHIVE_VALUE_PLACEHOLDERS =
-      "(?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)";
 
   // language=MySQL
   private static final String INSERT_ARCHIVE_SQL =
@@ -58,7 +45,9 @@ final class MysqlArchiveOperations implements ArchiveStore {
       INSERT INTO scheduler_job_archive (%s)
       VALUES %s
       """
-          .formatted(ARCHIVE_COLUMNS, ARCHIVE_VALUE_PLACEHOLDERS);
+          .formatted(
+              ArchiveParameterBinder.ARCHIVE_COLUMNS,
+              ArchiveParameterBinder.ARCHIVE_VALUE_PLACEHOLDERS);
 
   private final MysqlStoreContext ctx;
   private final MysqlJobRowMapper mapper;
@@ -85,45 +74,6 @@ final class MysqlArchiveOperations implements ArchiveStore {
     }
   }
 
-  private static int setArchiveParameters(Query query, ArchivedJobEntity archive, int parameter) {
-    query.setParameter(parameter++, UuidByteArrayConverter.toBytes(archive.getId()));
-    query.setParameter(parameter++, UuidByteArrayConverter.toBytes(archive.getOriginalJobId()));
-    query.setParameter(parameter++, archive.getFinalStatus().name());
-    query.setParameter(parameter++, archive.getJobType().name());
-    query.setParameter(parameter++, archive.getPriority().ordinal());
-    query.setParameter(parameter++, archive.getTotalAttempts());
-    query.setParameter(parameter++, archive.getMaxRetries());
-    query.setParameter(parameter++, archive.getBackoffPolicy().name());
-    query.setParameter(parameter++, archive.getBackoffParamMs());
-    query.setParameter(parameter++, archive.getTimeoutSec());
-    query.setParameter(parameter++, archive.getTargetClass());
-    query.setParameter(parameter++, archive.getMethodName());
-    query.setParameter(parameter++, archive.getBusinessKey());
-    query.setParameter(parameter++, archive.getCronExpr());
-    query.setParameter(parameter++, archive.getZoneId());
-    query.setParameter(parameter++, timestampOrNull(archive.getOriginalScheduledTime()));
-    query.setParameter(parameter++, timestampOrNull(archive.getOriginalCreatedAt()));
-    query.setParameter(parameter++, timestampOrNull(archive.getFirstExecutionTime()));
-    query.setParameter(parameter++, timestampOrNull(archive.getCompletionTime()));
-    query.setParameter(parameter++, archive.getTotalExecutionTimeMs());
-    query.setParameter(parameter++, archive.getQueueWaitMs());
-    query.setParameter(parameter++, timestampOrNull(archive.getArchivedAt()));
-    query.setParameter(parameter++, archive.getArchivedBy());
-    query.setParameter(parameter++, archive.getArchiveReason());
-    query.setParameter(parameter++, archive.getJobResult());
-    query.setParameter(parameter++, archive.getResultType());
-    query.setParameter(parameter++, archive.getFinalError());
-    query.setParameter(parameter++, archive.getPayloadSummary());
-    query.setParameter(parameter++, UuidByteArrayConverter.toBytes(archive.getDependedOn()));
-    query.setParameter(parameter++, UuidByteArrayConverter.toBytes(archive.getSupersededBy()));
-    query.setParameter(parameter++, archive.getTags());
-    return parameter;
-  }
-
-  private static Timestamp timestampOrNull(Instant instant) {
-    return instant == null ? null : Timestamp.from(instant);
-  }
-
   /**
    * Archives one terminal job in the caller's store transaction.
    *
@@ -136,7 +86,7 @@ final class MysqlArchiveOperations implements ArchiveStore {
       ArchivedJobEntity archive = ArchiveHelper.buildArchive(hydrated, reason, archivedBy);
       prepareArchive(archive);
       Query query = ctx.em().createNativeQuery(INSERT_ARCHIVE_SQL);
-      setArchiveParameters(query, archive, 1);
+      ArchiveParameterBinder.bind(query, archive, 1, UuidByteArrayConverter::toBytes);
       query.executeUpdate();
       return archive;
     } catch (RuntimeException e) {
@@ -173,7 +123,10 @@ final class MysqlArchiveOperations implements ArchiveStore {
       }
 
       String rows =
-          String.join(",", Collections.nCopies(archives.size(), ARCHIVE_VALUE_PLACEHOLDERS));
+          String.join(
+              ",",
+              Collections.nCopies(
+                  archives.size(), ArchiveParameterBinder.ARCHIVE_VALUE_PLACEHOLDERS));
       Query query =
           ctx.em()
               .createNativeQuery(
@@ -181,10 +134,11 @@ final class MysqlArchiveOperations implements ArchiveStore {
                   INSERT INTO scheduler_job_archive (%s)
                   VALUES %s
                   """
-                      .formatted(ARCHIVE_COLUMNS, rows));
+                      .formatted(ArchiveParameterBinder.ARCHIVE_COLUMNS, rows));
       int parameter = 1;
       for (ArchivedJobEntity archive : archives) {
-        parameter = setArchiveParameters(query, archive, parameter);
+        parameter =
+            ArchiveParameterBinder.bind(query, archive, parameter, UuidByteArrayConverter::toBytes);
       }
       query.executeUpdate();
       return archives.size();
@@ -256,7 +210,7 @@ final class MysqlArchiveOperations implements ArchiveStore {
     try {
       var searchQuery =
           ArchiveQuerySupport.buildFindArchivedJobsQuery(
-              ARCHIVE_COLUMNS, targetClass, businessKey, from, to, limit);
+              ArchiveParameterBinder.ARCHIVE_COLUMNS, targetClass, businessKey, from, to, limit);
       Query query = ctx.em().createNativeQuery(searchQuery.sql());
       ArchiveQuerySupport.bindParameters(query, searchQuery);
       @SuppressWarnings("unchecked")
